Restructured the animal.cpp file to be a entity/ directory. We can just continue to add however many entity we want and use this file to declare them all nice and simply.
