+++ Makefile
@@ -1,7 +1,7 @@
 
 CXX = g++
 CXXFLAGS = -std=c++20 -Wall -Wextra -Iinclude -g -Isrc 
-SRC = src/main.cpp src/glad.c src/graphics/Graphics.cpp src/ecs/components/renderobject.cpp src/engine/engine.cpp src/graphics/shaders.cpp src/ecs/spawner/spawner.cpp src/ecs/spawner/animals.cpp src/shape/shape.cpp
+SRC = src/main.cpp src/glad.c src/graphics/Graphics.cpp src/ecs/components/renderobject.cpp src/engine/engine.cpp src/graphics/shaders.cpp src/ecs/spawner/spawner.cpp src/ecs/spawner/entities/* src/shape/shape.cpp
 LIBS = -lglfw -ldl
 OUT = app
